+++ changelog/RELEASE-NOTES.md
@@ -1,2 +1,15 @@
+# NekoPara OneeSan Vol.
+
+**4 New Themes!!**
+
+- Maple (Light/Dark)
+- Cinnamon (Dark)
+- Azuki (Dark)
+
+![v18 Girls](https://doki.assets.unthrottled.io/misc/v18_girls_smol.png)
+
+Other Stuff
+---
+
 - Plugin is now supported on the `Code With Me` platform.
 - Added a convenient feature that allows you to set the console font family for all Doki Themes!